pyproject.toml
[project]
name = "resume-ranker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "bitsandbytes>=0.49.0",
    "einops>=0.8.1",
    "faiss-gpu-cu12>=1.13.0",
    "fastapi>=0.124.4",
    "loguru>=0.7.3",
    "opencv-python-headless>=4.12.0.88",
    "pdf2image>=1.17.0",
    "pydantic-settings>=2.12.0",
    "pytesseract>=0.3.13",
    "python-multipart>=0.0.20",
    "sentence-transformers>=5.2.0",
    "torch>=2.9.1",
    "torchaudio>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.3",
    "uvicorn>=0.38.0",
]
